refactor: Update BaseButton and BaseInput classes for better styling consistency
