+++ components/Base/BaseButton.vue
@@ -65,7 +65,7 @@
   :aria-disabled="disabled"
   :aria-label="title"
   :title="title"
-  class="flex items-center justify-center px-2 py-2 text-sm font-medium transition duration-300 ease-in-out transform border border-transparent rounded-lg shadow-lg md:px-4 focus:outline-none focus:ring-2 focus:ring-offset-2 hover:-translate-y-1 hover:scale-105"
+  class="flex items-center justify-center px-2 py-2 text-sm font-medium transition duration-300 ease-in-out transform border border-transparent rounded-lg shadow-lg md:px-4 focus:outline-none focus:ring-2 focus:ring-offset-2 hover:-translate-y-1 hover:scale-105 text-ellipsis"
   :class="[
     colorClasses,
     { 'inline-flex flex-row items-center space-x-1 md:space-x-2': $slots.icon },
